+++ chip8/src/chipvm/logic/instructions/Jumping.scala
@@ -7,31 +7,31 @@ import chipvm.logic._
  * Used to halt execution in case of an unknown instruction.
  */
 case class Nop() extends Instruction {
-  def execute(vm: ChipVMLogic): ChipVMLogic = vm.copy(pc = vm.pc-2)
+  def execute(vm: ChipVMLogic): ChipVMLogic = vm.copy(pc = vm.pc-UShort(2))
 }
 
 case class Return() extends Instruction {
   def execute(vm: ChipVMLogic): ChipVMLogic =
-    vm.copy(pc = vm.stack.head.toInt,
+    vm.copy(pc = vm.stack.head,
       stack = vm.stack.tail)
 }
 
 case class Jump(position: UShort) extends Instruction {
   def execute(vm: ChipVMLogic): ChipVMLogic =
-    vm.copy(pc = position.toInt)
+    vm.copy(pc = position)
 }
 
 case class JumpOffset(position: UShort) extends Instruction {
   def execute(vm: ChipVMLogic): ChipVMLogic = {
     val newPosition = position + vm.variableRegisters(0).toUShort
-    vm.copy(pc = newPosition.toInt)
+    vm.copy(pc = newPosition)
   }
 }
 
 case class CallSubroutine(position: UShort) extends Instruction {
   def execute(vm: ChipVMLogic): ChipVMLogic =
-    vm.copy(pc = position.toInt,
-      stack = vm.stack.prepended(UShort(vm.pc)))
+    vm.copy(pc = position,
+      stack = vm.stack.prepended(vm.pc))
 }
 
 // Conditional jumping instructions
@@ -39,7 +39,7 @@ case class CallSubroutine(position: UShort) extends Instruction {
 case class SkipValEqual(index: UByte, value: UByte) extends Instruction {
   def execute(vm: ChipVMLogic): ChipVMLogic = {
     vm.variableRegisters(index.toByte) match {
-      case `value` => vm.copy(pc = vm.pc + 2)
+      case `value` => vm.copy(pc = vm.pc + UShort(2))
       case _ => vm
     }
   }
@@ -49,7 +49,7 @@ case class SkipValNotEqual(index: UByte, value: UByte) extends Instruction {
   def execute(vm: ChipVMLogic): ChipVMLogic = {
     vm.variableRegisters(index.toByte) match {
       case `value` => vm
-      case _ => vm.copy(pc = vm.pc + 2)
+      case _ => vm.copy(pc = vm.pc + UShort(2))
     }
   }
 }
@@ -60,7 +60,7 @@ case class SkipRegisterEqual(index1: UByte, index2: UByte) extends Instruction {
     val value2 = vm.variableRegisters(index2.toByte)
 
     if (value1 == value2)
-      vm.copy(pc = vm.pc + 2)
+      vm.copy(pc = vm.pc + UShort(2))
     else
       vm
   }
@@ -72,7 +72,7 @@ case class SkipRegisterNotEqual(index1: UByte, index2: UByte) extends Instructio
     val value2 = vm.variableRegisters(index2.toByte)
 
     if (value1 != value2)
-      vm.copy(pc = vm.pc + 2)
+      vm.copy(pc = vm.pc + UShort(2))
     else
       vm
   }
