+++ chip8/src/chipvm/logic/instructions/Keypad.scala
@@ -1,6 +1,6 @@
 package chipvm.logic.instructions
 
-import chipvm.logic.{ChipVMLogic, UByte}
+import chipvm.logic.{ChipVMLogic, UByte, UShort}
 
 abstract class SkipKey(index: UByte, skipValue: Boolean) extends Instruction {
   def execute(vm: ChipVMLogic): ChipVMLogic = {
@@ -10,7 +10,7 @@ abstract class SkipKey(index: UByte, skipValue: Boolean) extends Instruction {
 
     val keepValue = !skipValue
     val pc = isKeyPressed match {
-      case `skipValue` => vm.pc + 2
+      case `skipValue` => vm.pc + UShort(2)
       case `keepValue` => vm.pc
     }
 
@@ -27,8 +27,8 @@ case class WaitForKey(index: UByte) extends Instruction {
       val firstKeyPressed = vm.pressedKeys.indexOf(true)
 
       firstKeyPressed match {
-        case -1 => vm.copy(pc = vm.pc - 2)
-        case key => vm.copy(pc = vm.pc - 2, waitForKeyIndex = Some(key))
+        case -1 => vm.copy(pc = vm.pc - UShort(2))
+        case key => vm.copy(pc = vm.pc - UShort(2), waitForKeyIndex = Some(key))
       }
     } else {
       val key = vm.waitForKeyIndex.get
@@ -36,7 +36,7 @@ case class WaitForKey(index: UByte) extends Instruction {
         val registers = vm.variableRegisters.updated(index.toInt, UByte(key))
         vm.copy(variableRegisters = registers, waitForKeyIndex = None)
       } else {
-        vm.copy(pc = vm.pc - 2)
+        vm.copy(pc = vm.pc - UShort(2))
       }
     }
   }
